refactor: place graphql queries inside graphql folder
